fix(types/containers): Properly split dashboards.ts into focused modules
🔧 Problem Resolution:
• Original dashboards.ts was still 520 lines (not properly divided)
• Configuration types were mixed with business logic types

✂️ Proper File Division:
• dashboards.ts: 412 lines → Business dashboard views only
• dashboard-types.ts: 370 lines → Configuration + widgets + session management

🛡️ Type Safety Improvements:
• Eliminated last Record<string, number> occurrence
• Added AlertThresholdType and AlertThresholdOverrides for type safety
• Maintained all interface exports and backward compatibility

📊 Architecture Results:
• ContainerArrivalDashboard, ManagerDashboardView, OperatorDashboardView, ClientDashboardView → dashboards.ts
• DashboardConfiguration, DashboardSession, AlertThresholdOverrides → dashboard-types.ts
• Clear separation between business logic and configuration concerns

🎯 Validation:
• TypeScript compilation: ✅ No errors
• All imports maintained in index.ts
• No breaking changes to existing API
