+++ types/containers/dashboards.ts
@@ -23,13 +23,6 @@ import type {
   ArrivalPerformanceMetrics
 } from './metrics';
 
-import type {
-  WidgetConfiguration,
-  DashboardFilters,
-  CachedDashboardData,
-  NavigationDetails
-} from './dashboard-types';
-
 // ============================================================================
 // Dashboard principal des arrivées
 // ============================================================================
@@ -417,105 +410,4 @@ export interface CostImpactAnalysis {
     best_in_class_benchmark: number;
     gap_analysis: string[];
   };
-}
-
-// ============================================================================
-// Configuration et session des dashboards
-// ============================================================================
-
-/**
- * Configuration personnalisée d'un dashboard
- */
-export interface DashboardConfiguration {
-  dashboard_id: string;
-  user_id: string;
-  dashboard_type: 'manager' | 'operator' | 'client' | 'analyst';
-  
-  // Layout et widgets
-  layout: {
-    widgets: Array<{
-      widget_id: string;
-      widget_type: 'kpi' | 'chart' | 'table' | 'alert' | 'map';
-      position: { x: number; y: number; width: number; height: number };
-      configuration: WidgetConfiguration;
-      is_visible: boolean;
-      refresh_interval_seconds?: number;
-    }>;
-    
-    theme: 'light' | 'dark' | 'auto';
-    layout_type: 'grid' | 'flex' | 'custom';
-  };
-  
-  // Filtres par défaut
-  default_filters: {
-    date_range: 'today' | 'week' | 'month' | 'quarter' | 'custom';
-    custom_date_from?: string;
-    custom_date_to?: string;
-    shipping_companies?: string[];
-    priority_levels?: string[];
-    client_filter?: string[];
-  };
-  
-  // Alertes et notifications
-  alert_preferences: {
-    real_time_alerts: boolean;
-    email_digest: 'none' | 'daily' | 'weekly';
-    push_notifications: boolean;
-    alert_sound: boolean;
-    alert_threshold_overrides?: Record<string, number>;
-  };
-  
-  // Personnalisation
-  display_preferences: {
-    timezone: string;
-    date_format: string;
-    number_format: string;
-    language: 'fr' | 'en' | 'es';
-    currency: string;
-  };
-  
-  created_at: string;
-  updated_at: string;
-  last_accessed: string;
-}
-
-/**
- * État de session d'un dashboard
- */
-export interface DashboardSession {
-  session_id: string;
-  user_id: string;
-  dashboard_id: string;
-  
-  // État actuel
-  current_filters: DashboardFilters;
-  selected_date_range: {
-    start_date: string;
-    end_date: string;
-  };
-  active_view: string;
-  zoom_level?: number;
-  
-  // Données en cache
-  cached_data?: CachedDashboardData;
-  cache_timestamp?: string;
-  cache_expiry?: string;
-  
-  // Historique de navigation
-  navigation_history: Array<{
-    timestamp: string;
-    action: 'filter_change' | 'view_change' | 'drill_down' | 'export';
-    details: NavigationDetails;
-  }>;
-  
-  // Métriques de session
-  session_start: string;
-  last_activity: string;
-  page_views: number;
-  actions_performed: number;
-  
-  // État de synchronisation
-  is_live_sync: boolean;
-  last_data_refresh: string;
-  pending_updates: number;
 }
